Progress on ocli compatibility
New selectChildrenNode that takes care of the ={path1, ..., pathn} syntax.
New parsing rules to differentiate different kind of object paths (as discussed on zulip)
to go with these rules, pathConvert becomes pathNode and takes a mode among STD, PHYSICAL or STRAY_DEV
Ignores duplicate errors and "orientation should be on payload" for now, to make debugging easier.
Adapt GetOCLIAttributes to work with new paths, the rule should now be that a controller function takes a full path as input, the preprocessing should be done in the ast.
Integers can now contain spaces, as it is sometimes the case in current ocli files.